+++ Source/GPU/Vulkan/States/Post/Post_Pipeline.hpp
@@ -41,7 +41,7 @@ struct Post_Pipeline
             .sType                      = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO,
             .pNext                      = nullptr,
             .flags                      = 0,
-            .stageCount                 = ArrayCount(shader.stageInfo),
+            .stageCount                 = (uint32_t) ArrayCount(shader.stageInfo),
             .pStages                    = shader.stageInfo,
             .pVertexInputState          = &vertexInput,
             .pInputAssemblyState        = &inputAssembly,
